refactor: simplify body handling in request/response message decoding
