+++ packages/standard-server-peer/src/codec.test.ts
@@ -1,5 +1,5 @@
 import type { StandardHeaders } from '@orpc/standard-server'
-import { decodeRequestMessage, decodeResponseMessage, encodeRequestMessage, encodeResponseMessage, isEventIteratorHeaders, MessageType } from './codec'
+import { decodeRequestMessage, decodeResponseMessage, encodeRequestMessage, encodeResponseMessage, MessageType } from './codec'
 
 const MB10Headers: StandardHeaders = {}
 
@@ -317,7 +317,7 @@ describe('encode/decode request message', () => {
           body: blob,
         })
 
-        expect(message).toBeInstanceOf(ArrayBuffer)
+        expect(message).toBeTypeOf('string')
 
         const [id, type, payload] = await decodeRequestMessage(message)
 
@@ -410,7 +410,7 @@ describe('encode/decode request message', () => {
           body: file,
         })
 
-        expect(message).toBeInstanceOf(ArrayBuffer)
+        expect(message).toBeTypeOf('string')
 
         const [id, type, payload] = await decodeRequestMessage(message)
 
@@ -834,23 +834,4 @@ describe('encode/decode response message', () => {
 
     expect(await (payload as any).body.text()).toBe(json)
   })
-})
-
-it('isEventIteratorHeaders', () => {
-  expect(isEventIteratorHeaders({
-    'content-type': 'text/event-stream',
-  })).toBe(true)
-
-  expect(isEventIteratorHeaders({
-    'content-type': 'text/event-stream',
-    'content-disposition': '',
-  })).toBe(false)
-
-  expect(isEventIteratorHeaders({
-    'content-type': 'text/plain',
-  })).toBe(false)
-
-  expect(isEventIteratorHeaders({
-    'content-disposition': 'attachment; filename="test.pdf"',
-  })).toBe(false)
 })
